# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApplyPlanningDraftRequest": {
        "properties": {
          "expected_project_revision": {
            "minimum": 0.0,
            "title": "Expected Project Revision",
            "type": "integer"
          },
          "expected_source_hash": {
            "maxLength": 64,
            "minLength": 64,
            "title": "Expected Source Hash",
            "type": "string"
          }
        },
        "required": [
          "expected_project_revision",
          "expected_source_hash"
        ],
        "title": "ApplyPlanningDraftRequest",
        "type": "object"
      },
      "ApplyPlanningDraftResponse": {
        "properties": {
          "analysis": {
            "$ref": "#/components/schemas/PlanningAnalysisResponse"
          },
          "applied": {
            "title": "Applied",
            "type": "boolean"
          },
          "project_revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Revision"
          },
          "snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectSnapshotResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "source_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Hash"
          }
        },
        "required": [
          "applied",
          "project_revision",
          "source_hash",
          "analysis",
          "snapshot"
        ],
        "title": "ApplyPlanningDraftResponse",
        "type": "object"
      },
...
      "CaptureProjectSnapshotRequest": {
        "properties": {
          "expected_planning_revision": {
            "minimum": 0.0,
            "title": "Expected Planning Revision",
            "type": "integer"
          },
          "expected_project_revision": {
            "minimum": 0.0,
            "title": "Expected Project Revision",
            "type": "integer"
          },
          "expected_source_hash": {
            "maxLength": 64,
            "minLength": 64,
            "title": "Expected Source Hash",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "required": [
          "expected_project_revision",
          "expected_source_hash",
          "expected_planning_revision"
        ],
        "title": "CaptureProjectSnapshotRequest",
        "type": "object"
      },
...
      "CreatePlanningDraftRequest": {
        "properties": {
          "base_project_revision": {
            "minimum": 0.0,
            "title": "Base Project Revision",
            "type": "integer"
          },
          "base_source_hash": {
            "maxLength": 64,
            "minLength": 64,
            "title": "Base Source Hash",
            "type": "string"
          },
          "changes": {
            "items": {
              "$ref": "#/components/schemas/DraftDocumentChange"
            },
            "minItems": 1,
            "title": "Changes",
            "type": "array"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "required": [
          "base_project_revision",
          "base_source_hash",
          "changes"
        ],
        "title": "CreatePlanningDraftRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlanningDraftResponse": {
        "properties": {
          "applied_revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied Revision"
          },
          "base_project_revision": {
            "title": "Base Project Revision",
            "type": "integer"
          },
          "base_source_hash": {
            "title": "Base Source Hash",
            "type": "string"
          },
          "candidate_documents": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Candidate Documents",
            "type": "array"
          },
          "candidate_source_hash": {
            "title": "Candidate Source Hash",
            "type": "string"
          },
          "changes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Changes",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "locale": {
            "title": "Locale",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "rspdl_version": {
            "title": "Rspdl Version",
            "type": "string"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "wire_schema_version": {
            "title": "Wire Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "project_id",
          "base_project_revision",
          "base_source_hash",
          "candidate_source_hash",
          "changes",
          "candidate_documents",
          "summary",
          "rspdl_version",
          "wire_schema_version",
          "locale",
          "result",
          "applied_revision",
          "created_at",
          "updated_at"
        ],
        "title": "PlanningDraftResponse",
        "type": "object"
      },
...
      "PlanningStateResponse": {
        "properties": {
          "decisions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Decisions",
            "type": "array"
          },
          "messages": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Messages",
            "type": "array"
          },
          "metadata": {
            "$ref": "#/components/schemas/PlanningMetadata"
          },
          "project_revision": {
            "description": "확정 RSPDL 원문 변경 묶음 버전",
            "title": "Project Revision",
            "type": "integer"
          },
          "proposals": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Proposals",
            "type": "array"
          },
          "revision": {
            "description": "대화·결정·메타데이터의 독립 optimistic revision",
            "title": "Revision",
            "type": "integer"
          },
          "source_hash": {
            "description": "현재 확정 원문의 canonical workspace hash",
            "title": "Source Hash",
            "type": "string"
          }
        },
        "required": [
          "revision",
          "project_revision",
          "source_hash",
          "messages",
          "decisions",
          "proposals",
          "metadata"
        ],
        "title": "PlanningStateResponse",
        "type": "object"
      },
...
      "RestoreProjectSnapshotRequest": {
        "properties": {
          "expected_planning_revision": {
            "minimum": 0.0,
            "title": "Expected Planning Revision",
            "type": "integer"
          },
          "expected_project_revision": {
            "minimum": 0.0,
            "title": "Expected Project Revision",
            "type": "integer"
          },
          "expected_source_hash": {
            "maxLength": 64,
            "minLength": 64,
            "title": "Expected Source Hash",
            "type": "string"
          }
        },
        "required": [
          "expected_project_revision",
          "expected_source_hash",
          "expected_planning_revision"
        ],
        "title": "RestoreProjectSnapshotRequest",
        "type": "object"
      },
...
      "UpdatePlanningStateRequest": {
        "properties": {
          "decisions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Decisions",
            "type": "array"
          },
          "expected_revision": {
            "minimum": 0.0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "messages": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Messages",
            "type": "array"
          },
          "metadata": {
            "$ref": "#/components/schemas/PlanningMetadata"
          },
          "proposals": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Proposals",
            "type": "array"
          }
        },
        "required": [
          "expected_revision"
        ],
        "title": "UpdatePlanningStateRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "RSPDL 저작·검증 워크스페이스",
    "title": "dahaze API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/planning/drafts/{draft_id}": {
      "get": {
        "operationId": "get_planning_draft",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Draft Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "dahaze_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dahaze Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanningDraftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Planning Draft",
        "tags": [
          "planning"
        ]
      }
    },
    "/api/planning/drafts/{draft_id}/apply": {
      "post": {
        "operationId": "apply_planning_draft",
        "parameters": [
          {
            "in": "path",
            "name": "draft_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Draft Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "dahaze_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dahaze Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyPlanningDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyPlanningDraftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Planning Draft",
        "tags": [
          "planning"
        ]
      }
    },
...
    "/api/projects/{project_id}/planning": {
      "get": {
        "operationId": "get_planning_state",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "dahaze_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dahaze Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanningStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Planning State",
        "tags": [
          "planning"
        ]
      },
      "put": {
        "operationId": "update_planning_state",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "dahaze_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dahaze Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanningStateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanningStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Planning State",
        "tags": [
          "planning"
        ]
      }
    },
    "/api/projects/{project_id}/planning/drafts": {
      "get": {
        "operationId": "list_planning_drafts",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "dahaze_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dahaze Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlanningDraftResponse"
                  },
                  "title": "Response List Planning Drafts",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Planning Drafts",
        "tags": [
          "planning"
        ]
      },
      "post": {
        "operationId": "create_planning_draft",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "dahaze_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dahaze Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanningDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanningDraftResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Planning Draft",
        "tags": [
          "planning"
        ]
      }
    },
    "/api/projects/{project_id}/planning/handoff": {
      "get": {
        "operationId": "get_project_handoff",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "revision",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Revision",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "dahaze_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dahaze Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project Handoff",
        "tags": [
          "planning"
        ]
      }
    },
    "/api/projects/{project_id}/planning/snapshots": {
      "get": {
        "operationId": "list_project_snapshots",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "dahaze_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dahaze Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectSnapshotResponse"
                  },
                  "title": "Response List Project Snapshots",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Project Snapshots",
        "tags": [
          "planning"
        ]
      },
      "post": {
        "operationId": "capture_project_snapshot",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "dahaze_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dahaze Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptureProjectSnapshotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Capture Project Snapshot",
        "tags": [
          "planning"
        ]
      }
    },
    "/api/projects/{project_id}/planning/snapshots/{revision}/restore": {
      "post": {
        "operationId": "restore_project_snapshot",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision",
            "required": true,
            "schema": {
              "title": "Revision",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "dahaze_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Dahaze Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreProjectSnapshotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Project Snapshot",
        "tags": [
          "planning"
        ]
      }
    },
...
  }
}